fix: 전체 코드에 리팩터링 적용
