+++ src/main/kotlin/blackjack/model/participant/Player.kt
@@ -1,28 +1,19 @@
 package blackjack.model.participant
 
 import blackjack.model.deck.Card
-import blackjack.model.deck.HandCards
-import blackjack.model.participant.state.Blackjack
-import blackjack.model.participant.state.Bust
-import blackjack.model.participant.state.Normal
-import blackjack.model.participant.state.ParticipantState
+import blackjack.model.participant.state.Finish
+import blackjack.model.participant.state.Gaming
 
-class Player(val name: String) : GameParticipant(HandCards()) {
-    override fun add(cards: List<Card>): Boolean =
-        if (!isBust()) {
-            handCards.add(cards)
-            true
-        } else {
-            false
+class Player(val name: String) : GameParticipant() {
+    fun playTurn(
+        cards: (Int) -> List<Card>,
+        isHit: (String) -> Boolean,
+        showResult: (Player) -> Unit,
+    ): Finish {
+        while (handCards.state is Gaming) {
+            handCards.playTurn(isHit(name), cards)
+            showResult(this)
         }
-
-    fun getState(): ParticipantState {
-        return if (handCards.isBlackjackCard()) {
-            Blackjack()
-        } else if (isBust()) {
-            Bust()
-        } else {
-            Normal(getScore())
-        }
+        return handCards.state as Finish
     }
 }
