--- src/main/kotlin/blackjack/model/participant/state/Blackjack.kt
@@ -1,13 +1,0 @@
-package blackjack.model.participant.state
-
-import blackjack.model.participant.CompetitionResult
-
-class Blackjack : ParticipantState {
-    override fun getResult(opponentScore: Int): CompetitionResult {
-        return if (opponentScore == 21) {
-            CompetitionResult.SAME
-        } else {
-            CompetitionResult.WIN
-        }
-    }
-}
