--- src/main/kotlin/blackjack/model/participant/state/Bust.kt
@@ -1,9 +1,0 @@
-package blackjack.model.participant.state
-
-import blackjack.model.participant.CompetitionResult
-
-class Bust : ParticipantState {
-    override fun getResult(opponentScore: Int): CompetitionResult {
-        return CompetitionResult.LOSE
-    }
-}
