--- src/main/kotlin/blackjack/model/participant/state/ParticipantState.kt
@@ -1,7 +1,0 @@
-package blackjack.model.participant.state
-
-import blackjack.model.participant.CompetitionResult
-
-fun interface ParticipantState {
-    fun getResult(opponentScore: Int): CompetitionResult
-}
